# docs/swagger.yaml
definitions:
...
  dto.CreateEtsyShipmentRequest:
    properties:
      carrier_code:
        description: yanwen, wanbang
        type: string
      order_id:
        type: integer
      service_code:
        description: 服务代码
        type: string
      tracking_number:
        description: 可选，手动填写
        type: string
      weight:
        description: 重量 KG
        type: number
      weight_unit:
        description: KG, LB
        type: string
    required:
    - carrier_code
    - order_id
    type: object
  dto.CreateLabelRequest:
    properties:
      carrier_code:
        type: string
      order_id:
        type: integer
      service_code:
        type: string
    required:
    - carrier_code
    - order_id
    - service_code
    type: object
...
info:
  contact: {}
paths:
...
  /api/orders/{id}/shipment:
    get:
      description: 获取指定订单的发货记录详情
      parameters:
      - description: 订单ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"data": dto.ShipmentVO}'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: ID格式错误
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 发货记录不存在
          schema:
            additionalProperties:
              type: string
            type: object
        "503":
          description: 发货服务未配置
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 获取订单发货信息
      tags:
      - Order (订单管理)
...
  /api/orders/{order_id}/shipment:
    get:
      description: 获取指定订单的发货记录
      parameters:
      - description: 订单ID
        in: path
        name: order_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"data": dto.EtsyShipmentResponse}'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: ID格式错误
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 发货记录不存在
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 根据订单ID获取发货信息
      tags:
      - Shipment (发货管理)
...
  /api/shipments:
    get:
      description: 分页查询发货记录，支持多条件筛选
      parameters:
      - description: 订单ID
        in: query
        name: order_id
        type: integer
      - description: 物流商代码
        in: query
        name: carrier_code
        type: string
      - description: 发货状态
        in: query
        name: status
        type: string
      - description: 运单号
        in: query
        name: tracking_number
        type: string
      - description: 是否已同步Etsy
        in: query
        name: etsy_synced
        type: boolean
      - description: 开始日期 (YYYY-MM-DD)
        in: query
        name: start_date
        type: string
      - description: 结束日期 (YYYY-MM-DD)
        in: query
        name: end_date
        type: string
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 20
        description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"data": dto.EtsyShipmentListResponse}'
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器错误
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 获取发货列表
      tags:
      - Shipment (发货管理)
    post:
      consumes:
      - application/json
      description: 为订单创建发货记录，填写物流商和运单号
      parameters:
      - description: 发货参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateEtsyShipmentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: '{"data": dto.EtsyShipmentResponse, "message": "发货成功"}'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 参数错误
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 创建发货记录
      tags:
      - Shipment (发货管理)
  /api/shipments/{id}:
    get:
      description: 根据发货记录ID获取详细信息
      parameters:
      - description: 发货记录ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"data": dto.EtsyShipmentResponse}'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: ID格式错误
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 发货记录不存在
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 获取发货详情
      tags:
      - Shipment (发货管理)
  /api/shipments/{id}/refresh-tracking:
    post:
      description: 手动刷新指定发货记录的物流轨迹
      parameters:
      - description: 发货记录ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"data": dto.EtsyShipmentResponse, "message": "物流信息已刷新"}'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 刷新失败
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 刷新物流跟踪状态
      tags:
      - Shipment (发货管理)
  /api/shipments/{id}/sync-etsy:
    post:
      description: 将发货信息同步到 Etsy 平台，更新订单的发货状态
      parameters:
      - description: 发货记录ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"message": "已同步到 Etsy"}'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 同步失败
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 同步发货信息到 Etsy
      tags:
      - Shipment (发货管理)
...
  /api/shipments/with-label:
    post:
      consumes:
      - application/json
      description: 为订单创建发货，自动生成物流面单
      parameters:
      - description: 面单参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateLabelRequest'
      produces:
      - application/json
      responses:
        "201":
          description: '{"data": dto.EtsyShipmentResponse, "message": "发货成功，面单已生成"}'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 参数错误
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 创建发货并生成面单
      tags:
      - Shipment (发货管理)
...
  /api/webhooks/karrio/tracking:
    post:
      consumes:
      - application/json
      description: 接收 Karrio 推送的物流状态更新事件
      parameters:
      - description: Webhook 载荷
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: '{"message": "processed"}'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 处理失败
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 处理 Karrio 物流回调
      tags:
      - Shipment (发货管理)
...
swagger: "2.0"